fix: Restore meeting state after page refresh (#377)
Move meeting state restoration from NoteCapture to MainContent so it
runs regardless of which tab is active. Previously, refreshing while
on any tab except Capture would orphan an active meeting because
get_meeting_state was never sent.

Changes:
- App.tsx: Send select_vault and get_meeting_state on mount and reconnect
- NoteCapture.tsx: Remove redundant vault/meeting restoration logic
